pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pandemic-rl"
version = "0.1.0"
description = "Pandemic RL OpenEnv Hackathon Project"
authors = [{ name = "Dharmveer" }]

dependencies = [
    "numpy>=1.21.0",
    "requests>=2.25.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "openenv-core>=0.2.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
include = ["pandemic_rl*", "env*", "server*"]